LayoutTests/imported/w3c/web-platform-tests/css/cssom/caretPositionFromPoint-with-transformation-expected.txt

CONSOLE MESSAGE: TypeError: frameDoc.caretPositionFromPoint is not a function. (In 'frameDoc.caretPositionFromPoint(
          ...elementCenter(frameDoc.querySelector("h1"))
        )', 'frameDoc.caretPositionFromPoint' is undefined)
CONSOLE MESSAGE: TypeError: frameDoc.caretPositionFromPoint is not a function. (In 'frameDoc.caretPositionFromPoint(
          ...elementCenter(frameDoc.querySelector("h1"))
        )', 'frameDoc.caretPositionFromPoint' is undefined)



Harness Error (FAIL), message = TypeError: frameDoc.caretPositionFromPoint is not a function. (In 'frameDoc.caretPositionFromPoint(
          ...elementCenter(frameDoc.querySelector("h1"))
        )', 'frameDoc.caretPositionFromPoint' is undefined)

TIMEOUT iframe's with equal content should report the same caret offset Test timed out

Harness Error (FAIL), message = TypeError: frameDoc.caretPositionFromPoint is not a function. (In 'frameDoc.caretPositionFromPoint(
          ...elementCenter(frameDoc.querySelector("h1"))
        )', 'frameDoc.caretPositionFromPoint' is undefined)

TIMEOUT iframe's with equal content should report the same caret offset Test timed out
